airflow/requirements.txt
absl-py==0.15.0
appdirs==1.4.4
astunparse==1.6.3
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
cssselect==1.1.0
cycler==0.11.0
DateTime==4.3
fake-useragent==0.1.11
filelock==3.3.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gdown==4.2.0
google-api-core==2.0.0
google-api-python-client==2.17.0
google-auth==2.0.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
google-pasta==0.2.0
grpcio==1.34.1
httplib2==0.20.4
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.6.4
Markdown==3.3.4
matplotlib==3.4.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
parse==1.19.0
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.2.2
pymongo==3.12.1
pyparsing==2.4.2
pyppeteer==0.2.6
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.26.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
searchtweets-v2==1.1.1
six==1.15.0
soupsieve==2.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.1
tensorflow-estimator==2.5.0
termcolor==1.1.0
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.7
w3lib==1.22.0
websockets==9.1
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
zope.interface==5.4.0
